Fix validation and code style

+++ passenger-service/src/main/java/com/modsen/service/impl/PassengerServiceImpl.java
@@ -1,5 +1,7 @@
 package com.modsen.service.impl;
 
+import com.modsen.constants.PassengerServiceConstants;
+import com.modsen.dto.PassengerListResponse;
 import com.modsen.dto.PassengerRequest;
 import com.modsen.dto.PassengerResponse;
 import com.modsen.exception.PassengerNotFoundException;
@@ -10,6 +12,7 @@ import com.modsen.repository.PassengerRepository;
 import com.modsen.service.PassengerService;
 import com.modsen.util.PageRequestFactory;
 import lombok.RequiredArgsConstructor;
+import org.springframework.dao.DuplicateKeyException;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
@@ -21,33 +24,83 @@ public class PassengerServiceImpl implements PassengerService {
     private final PassengerRepository passengerRepository;
 
     @Override
-    public List<PassengerResponse> getAllPassenger(PageSetting pageSetting) {
+    public PassengerListResponse getAllPassenger(PageSetting pageSetting) {
         Pageable pageable = PageRequestFactory.buildPageRequest(pageSetting);
-        passengerRepository.findAll(pageable).forEach(System.out::println);
-        List<Passenger> passengerList = passengerRepository.findAll(pageable).toList();
-        return PassengerMapper.MAPPER.mapToListOfPassengerResponse(passengerList);
+        List<Passenger> passengers = passengerRepository.findAll(pageable)
+                .toList();
+        List<PassengerResponse> passengerResponses = PassengerMapper.MAPPER_INSTANCE
+                .mapToListOfPassengerResponse(passengers);
+
+        return PassengerListResponse.builder()
+                .passengers(passengerResponses)
+                .totalPassengerCount(passengerResponses.size())
+                .build();
     }
 
     @Override
-    public void createPassenger(PassengerRequest passengerRequest) {
-        Passenger newPassenger = PassengerMapper.MAPPER.mapToPassenger(passengerRequest);
-        passengerRepository.save(newPassenger);
+    public PassengerResponse createPassenger(PassengerRequest passengerRequest) {
+        validatePassengerRequest(passengerRequest);
+        Passenger newPassenger = PassengerMapper.MAPPER_INSTANCE
+                .mapToPassenger(passengerRequest);
+
+        return PassengerMapper.MAPPER_INSTANCE.mapToPassengerResponse(
+                passengerRepository.save(newPassenger)
+        );
     }
 
     @Override
-    public PassengerResponse getPassengerById(Integer id) {
-        Passenger foundPassenger = passengerRepository.findById(id).orElseThrow(() -> new PassengerNotFoundException(id));
-        return PassengerMapper.MAPPER.mapToPassengerResponse(foundPassenger);
+    public PassengerResponse getPassengerById(long id) {
+        Passenger foundPassenger = passengerRepository.findById(id)
+                .orElseThrow(() -> new PassengerNotFoundException(id));
+
+        return PassengerMapper.MAPPER_INSTANCE
+                .mapToPassengerResponse(foundPassenger);
     }
 
     @Override
-    public void updatePassenger(Integer id, PassengerRequest passengerRequest) {
-        Passenger updatedPassenger = PassengerMapper.MAPPER.mapToPassenger(passengerRequest);
+    public PassengerResponse updatePassenger(long id, PassengerRequest passengerRequest) {
+        if(!passengerRepository.existsById(id)){
+            throw new PassengerNotFoundException(id);
+        }
+
+        validatePassengerRequest(passengerRequest);
+        Passenger updatedPassenger = PassengerMapper.MAPPER_INSTANCE
+                .mapToPassenger(passengerRequest);
         updatedPassenger.setId(id);
-        passengerRepository.save(updatedPassenger);
+
+        return PassengerMapper.MAPPER_INSTANCE
+                .mapToPassengerResponse(
+                    passengerRepository.save(updatedPassenger)
+        );
     }
 
-    public void deletePassenger(Integer id) {
-        passengerRepository.deleteById(id);
+    private void validatePassengerRequest(PassengerRequest passengerRequest) {
+        if(passengerRepository.existsByEmail(passengerRequest.getEmail())) {
+            throw new DuplicateKeyException(
+                    String.format(
+                            PassengerServiceConstants.Errors.Message.DUPLICATE_USER_WITH_EMAIL,
+                            passengerRequest.getEmail()
+                    )
+            );
+        }
+
+        if(passengerRepository.existsByPhone(passengerRequest.getPhone())) {
+            throw new DuplicateKeyException(
+                    String.format(
+                            PassengerServiceConstants.Errors.Message.DUPLICATE_USER_WITH_PHONE,
+                            passengerRequest.getPhone()
+                    )
+            );
+        }
     }
+
+    public void deletePassenger(long id) {
+        if(passengerRepository.existsById(id)) {
+            passengerRepository.deleteById(id);
+        } else {
+            throw new PassengerNotFoundException(id);
+        }
+    }
+
+
 }
